requirements.txt
discord.py>=2.3.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.5.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiosqlite
jinja2